+++ Classes/VTAcknowledgementsViewController.m
@@ -82,7 +82,7 @@ static const CGFloat VTFooterBottomMargin = 20;
     return [[self.class alloc] initWithAcknowledgementsPlistPath:path];
 }
 
-- (instancetype)initWithAcknowledgementsPlistPath:(NSString *)acknowledgementsPlistPath
+- (instancetype)initWithPath:(NSString *)acknowledgementsPlistPath
 {
     self = [super initWithStyle:UITableViewStyleGrouped];
     if (self) {
@@ -92,10 +92,20 @@ static const CGFloat VTFooterBottomMargin = 20;
     return self;
 }
 
-- (nullable instancetype)initWithAcknowledgementsFileNamed:(nullable NSString *)acknowledgementsFileName
+- (instancetype)initWithAcknowledgementsPlistPath:(NSString *)acknowledgementsPlistPath
+{
+    return [self initWithPath:acknowledgementsPlistPath];
+}
+
+- (nullable instancetype)initWithFileNamed:(nonnull NSString *)acknowledgementsFileName
 {
     NSString *path = [[NSBundle mainBundle] pathForResource:acknowledgementsFileName ofType:@"plist"];
-    return [self initWithAcknowledgementsPlistPath:path];
+    return [self initWithPath:path];
+}
+
+- (nullable instancetype)initWithAcknowledgementsFileNamed:(nullable NSString *)acknowledgementsFileName
+{
+    return [self initWithFileNamed:acknowledgementsFileName];
 }
 
 - (void)awakeFromNib